Fix #17932 - [Bug] Fix Transactional removal cache is full

+++ becpg-core/src/main/java/fr/becpg/repo/repository/impl/AlfrescoRepositoryImpl.java
@@ -182,13 +182,17 @@ public class AlfrescoRepositoryImpl<T extends RepositoryEntity>
 	
 	@Override
 	public void onAddAspect(NodeRef nodeRef, QName aspectTypeQName) {
-		purgeCache(nodeRef);
+		if(nodeService.exists(nodeRef) && entityDictionaryService.isSubClass(nodeService.getType(nodeRef), BeCPGModel.TYPE_ENTITY_V2)) {
+			purgeCache(nodeRef);
+		}
 		
 	}
 	
 	@Override
 	public void onRemoveAspect(NodeRef nodeRef, QName aspectTypeQName) {
-		purgeCache(nodeRef);
+		if(nodeService.exists(nodeRef) && entityDictionaryService.isSubClass(nodeService.getType(nodeRef), BeCPGModel.TYPE_ENTITY_V2)) {
+			purgeCache(nodeRef);
+		}
 	}
 	
 	@Override
@@ -202,13 +206,15 @@ public class AlfrescoRepositoryImpl<T extends RepositoryEntity>
 	}
 
 	private void purgeCache(NodeRef nodeRef) {
-		if (logger.isDebugEnabled()) {
+		
 			if (nodeService.exists(nodeRef)) {
 				QName type = nodeService.getType(nodeRef);
 				Class<T> entityClass = repositoryEntityDefReader.getEntityClass(type);
 				if (entityClass != null) {
 					if(entityClass.isAnnotationPresent(AlfCacheable.class) ) {
-						logger.info("Clear cache of:" + nodeRef + " - " + nodeService.getProperty(nodeRef, ContentModel.PROP_NAME));
+						if (logger.isDebugEnabled()) {
+							logger.info("Clear cache of:" + nodeRef + " - " + nodeService.getProperty(nodeRef, ContentModel.PROP_NAME));
+						}
 						if (entityClass.getAnnotation(AlfCacheable.class).isCharact()) {
 							charactCache.remove(nodeRef);
 						} else {
@@ -217,13 +223,12 @@ public class AlfrescoRepositoryImpl<T extends RepositoryEntity>
 					}
 				}
 			} else {
-				logger.info("Clear cache of:" + nodeRef + " - deleted");
+				if (logger.isDebugEnabled()) {
+					logger.info("Clear cache of:" + nodeRef + " - deleted");
+				}
 				charactCache.remove(nodeRef);
 				cache.remove(nodeRef);
 			}
-		}
-	
-	
 	}
 
 	/** {@inheritDoc} */
